% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normc.R
\name{normc}
\alias{normc}
\title{Normaliz the columns of x to a length of 1.}
\usage{
normc(x)
}
\arguments{
\item{x}{n*p matrix}
}
\value{
xn normalized result
}
\description{
Normaliz the columns of x to a length of 1.
}
\examples{
#ex1.
m <- matrix(1:4,2,2,byrow=TRUE)
normc(m)
#ex2.
n <- matrix(rnorm(100,10,1),10,10)
normc(n)
}
